+++ Remote/src/Dap.Remote.FSharpData/RELEASE_NOTES.md
@@ -1,2 +1,5 @@
+## 0.6.1
+* Http.encodeXXX helpers
+
 ## 0.6.0
 * Http to get data from remote services